+++ src/pas/plugins/imio/browser/view.py
@@ -211,6 +211,10 @@ class AuthenticView(BrowserView):
             self.request,
         )
 
+    def next(self):
+        """ Used to login page view """
+        return self.request.form.get("next_url", "/")
+
     def __call__(self):
         cfg = authentic_cfg()
         if cfg is None:
@@ -245,7 +249,8 @@ class AuthenticView(BrowserView):
         if six.PY2 and isinstance(secret, six.text_type):
             secret = secret.encode("utf8")
         auth = Authomatic(cfg, secret=secret)
-        result = auth.login(ZopeRequestAdapter(self), self.provider)
+        zope_request_adapter = ZopeRequestAdapter(self)
+        result = auth.login(zope_request_adapter, self.provider)
         if not result:
             logger.info("return from view")
             # let authomatic do its work
@@ -262,12 +267,15 @@ class AuthenticView(BrowserView):
         else:
             # now we delegate to PAS plugin in order to login
             self._remember_identity(result, provider_name)
+            redirect_url = self.context.absolute_url()
             if api.env.plone_version() < "5.2":
-                self.request.response.redirect(
-                    "{0}/login_success".format(self.context.absolute_url())
-                )
+                self.request.response.redirect("{0}/login_success".format(redirect_url))
             else:
-                self.request.response.redirect(self.context.absolute_url())
+                state = result.provider.params.get("state")
+                if state:
+                    decoded_state = result.provider.decode_state(state)
+                    redirect_url = "{0}{1}".format(redirect_url, decoded_state)
+                self.request.response.redirect(redirect_url)
         return "redirecting"
 
     @property