+++ src/pas/plugins/imio/login.py
@@ -7,6 +7,10 @@ class ImioLoginFormView(BrowserView):
     def __call__(self):
         """Redirect login to authentic"""
         response = self.request.response
+        next_url = self.request.get("HTTP_REFERER", self.context.absolute_url())
+        next_url = next_url.replace(api.portal.get().absolute_url(), "")
         response.redirect(
-            "{0}/authentic-handler".format(api.portal.get().absolute_url())
+            "{0}/authentic-handler/?next_url={1}".format(
+                api.portal.get().absolute_url(), next_url
+            )
         )