+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0b9 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Use state / user_state to redirect to page which apply SSO.
+  [bsuttor]
 
 
 1.0b8 (2020-02-21)